Add plausi check for phase configuration

+++ core/helper.go
@@ -16,7 +16,7 @@ var (
 )
 
 // powerToCurrent is a helper function to convert power to per-phase current
-func powerToCurrent(power float64, phases int64) float64 {
+func powerToCurrent(power float64, phases int) float64 {
 	if Voltage == 0 {
 		panic("Voltage is not set")
 	}
